fix code ss2
