+++ reactjs/examples/src/components/Session_2/Footer/columns.ts
@@ -1,0 +1,44 @@
+const columns = [
+  {
+    title: "Avatar",
+    dataIndex: "Avatar",
+    key: "Avatar",
+  },
+  {
+    title: "Name",
+    dataIndex: "Ten",
+    key: "Name",
+  },
+  {
+    title: "Phone",
+    dataIndex: "SDT",
+    key: "Phone",
+  },
+  {
+    title: "Room",
+    dataIndex: "LoaiPhong",
+    key: "Room",
+  },
+  {
+    title: "CheckIn",
+    dataIndex: "Checkin",
+    key: "CheckIn",
+  },
+  {
+    title: "CheckOut",
+    dataIndex: "Checkout",
+    key: "CheckOut",
+  },
+  {
+    title: "Status",
+    dataIndex: "TrangThai",
+    key: "Status",
+  },
+  {
+    title: "Action",
+    dataIndex: "HanhDong",
+    key: "Action",
+  },
+];
+
+export default columns;
