+++ reactjs/examples/src/components/Session_2/Footer/index.tsx
@@ -3,60 +3,16 @@ import { bookingCollection } from "../../../firebase/controller";
 import { DocumentData, QuerySnapshot, onSnapshot } from "firebase/firestore";
 import { BookingList } from "../types";
 import { Table } from "antd";
+import columns from "./columns";
 
 function Footer() {
-  const dataSource = [
-    {
-      key: "Action",
-      Action: "...",
-    },
-  ];
-  const columns = [
-    {
-      title: "Avatar",
-      dataIndex: "Avatar",
-      key: "Avatar",
-    },
-    {
-      title: "Name",
-      dataIndex: "Ten",
-      key: "Name",
-    },
-    {
-      title: "Phone",
-      dataIndex: "SDT",
-      key: "Phone",
-    },
-    {
-      title: "Room",
-      dataIndex: "LoaiPhong",
-      key: "Room",
-    },
-    {
-      title: "CheckIn",
-      dataIndex: "Checkin",
-      key: "CheckIn",
-    },
-    {
-      title: "CheckOut",
-      dataIndex: "Checkout",
-      key: "CheckOut",
-    },
-    {
-      title: "Status",
-      dataIndex: "TrangThai",
-      key: "Status",
-    },
-    {
-      title: "Action",
-      dataIndex: "HanhDong",
-      key: "Action",
-    },
-  ];
   const [book, setBook] = useState<BookingList[]>([]);
-  useEffect(
-    () =>
-      onSnapshot(bookingCollection, (snapshot: QuerySnapshot<DocumentData>) => {
+  const [unsubscribe, setUnsubscribe] = useState<(() => void) | null>(null);
+
+  useEffect(() => {
+    const unsubscribeRef = onSnapshot(
+      bookingCollection,
+      (snapshot: QuerySnapshot<DocumentData>) => {
         setBook(
           snapshot.docs.map((doc) => {
             return {
@@ -65,9 +21,15 @@ function Footer() {
             };
           })
         );
-      }),
-    []
-  );
+      }
+    );
+    setUnsubscribe(() => unsubscribeRef);
+    return () => {
+      if (unsubscribe) {
+        unsubscribe();
+      }
+    };
+  }, []);
   console.log(book, "BookingList");
   return (
     <div className="Footer">
